+++ helpers/index.js
@@ -1,5 +1,2 @@
-export { default as authenticate } from "./authenticate.js";
-export { default as HttpError } from "./HttpError.js";
-export { default as validateBody } from "./validateBody.js";
-export { default as validateId } from "./validateId.js";
 export { default as ctrlWrapper } from "./ctrlWrapper.js";
+export { default as HttpError } from "./HttpError.js";
