+++ middlewares/index.js
@@ -1,0 +1,3 @@
+export { default as authenticate } from "./authenticate.js";
+export { default as validateBody } from "./validateBody.js";
+export { default as validateId } from "./validateId.js";
