+++ middlewares/validateBody.js
@@ -1,4 +1,4 @@
-import { HttpError } from "./index.js";
+import { HttpError } from "../helpers/index.js";
 
 const validateBody = (schema) => {
   const func = (req, _, next) => {
